+++ inc/authldap.class.php
@@ -1779,7 +1779,7 @@ class AuthLDAP extends CommonDBTM {
                // -> renaming case
                if ($userfound) {
                   //Get user in DB with this dn
-                  if (!$tmpuser->getFromDBByDn(Toolbox::addslashes_deep($user['user_dn']))) {
+                  if (!$tmpuser->getFromDBByDn($user['user_dn'])) {
                      //This should never happened
                      //If a user_dn is present more than one time in database
                      //Just skip user synchronization to avoid errors
@@ -1824,7 +1824,7 @@ class AuthLDAP extends CommonDBTM {
 
          foreach ($diff as $user) {
             //If user dn exists in DB, it means that user login field has changed
-            if (!$tmpuser->getFromDBByDn(toolbox::addslashes_deep($user_infos[$user]["user_dn"]))) {
+            if (!$tmpuser->getFromDBByDn($user_infos[$user]["user_dn"])) {
                $entry  = ["user"      => $user_infos[$user][$config_ldap->fields['login_field']],
                           "timestamp" => $user_infos[$user]["timestamp"],
                           "date_sync" => Dropdown::EMPTY_VALUE];
@@ -2191,7 +2191,7 @@ class AuthLDAP extends CommonDBTM {
                               'SELECT' => ['ldap_value'],
                               'FROM'   => 'glpi_groups',
                               'WHERE'  => [
-                                 'ldap_group_dn' => Toolbox::addslashes_deep($ou)
+                                 'ldap_group_dn' => $ou
                               ]
                            ]);
 
@@ -2377,7 +2377,7 @@ class AuthLDAP extends CommonDBTM {
                $user    = new User();
 
                //Get information from LDAP
-               if ($user->getFromLDAP($ds, $config_ldap->fields, $user_dn, addslashes($login),
+               if ($user->getFromLDAP($ds, $config_ldap->fields, $user_dn, $login,
                                     ($action == self::ACTION_IMPORT))) {
                   // Add the auth method
                   // Force date sync
@@ -2463,14 +2463,14 @@ class AuthLDAP extends CommonDBTM {
          $group_infos = self::getGroupByDn($ds, stripslashes($group_dn));
          $group       = new Group();
          if ($options['type'] == "groups") {
-            return $group->add(["name"          => addslashes($group_infos["cn"][0]),
-                                "ldap_group_dn" => addslashes($group_infos["dn"]),
+            return $group->add(["name"          => $group_infos["cn"][0],
+                                "ldap_group_dn" => $group_infos["dn"],
                                 "entities_id"   => $options['entities_id'],
                                 "is_recursive"  => $options['is_recursive']]);
          }
-         return $group->add(["name"         => addslashes($group_infos["cn"][0]),
+         return $group->add(["name"         => $group_infos["cn"][0],
                              "ldap_field"   => $config_ldap->fields["group_field"],
-                             "ldap_value"   => addslashes($group_infos["dn"]),
+                             "ldap_value"   => $group_infos["dn"],
                              "entities_id"  => $options['entities_id'],
                              "is_recursive" => $options['is_recursive']]);
       }
@@ -2679,7 +2679,7 @@ class AuthLDAP extends CommonDBTM {
 
       if ($user_dn) {
          $auth->auth_succeded            = true;
-         if ($auth->user->getFromDBbyDn(toolbox::addslashes_deep($user_dn))) {
+         if ($auth->user->getFromDBbyDn($user_dn)) {
             //There's already an existing user in DB with the same DN but its login field has changed
             $auth->user->fields['name'] = $login;
             $auth->user_present         = true;
@@ -3605,7 +3605,7 @@ class AuthLDAP extends CommonDBTM {
          return $user;
       }
 
-      if ($user->getFromDBbyNameAndAuth($DB->escape($name), Auth::LDAP, $authldaps_id)) {
+      if ($user->getFromDBbyNameAndAuth($name, Auth::LDAP, $authldaps_id)) {
          return $user;
       }