+++ inc/cartridge.class.php
@@ -235,7 +235,7 @@ class Cartridge extends CommonDBChild {
             'id' => $input['id']
          ]
       );
-      if ($result && ($DB->affected_rows() > 0)) {
+      if ($result && ($result->rowCount() > 0)) {
          return true;
       }
       return false;
@@ -281,7 +281,7 @@ class Cartridge extends CommonDBChild {
                'date_use'  => null
             ]
          );
-         if ($result && ($DB->affected_rows() > 0)) {
+         if ($result && ($result->rowCount() > 0)) {
             $changes[0] = '0';
             $changes[1] = '';
             $changes[2] = __('Installing a cartridge');
@@ -321,7 +321,7 @@ class Cartridge extends CommonDBChild {
                    WHERE `id`='$ID'";
 
          if ($result = $DB->query($query)
-             && ($DB->affected_rows() > 0)) {
+             && ($result->rowCount() > 0)) {
             $changes[0] = '0';
             $changes[1] = '';
             $changes[2] = __('Uninstalling a cartridge');