+++ inc/crontasklog.class.php
@@ -58,13 +58,14 @@ class CronTaskLog extends CommonDBTM{
 
       $secs      = $days * DAY_TIMESTAMP;
 
+      //TODO: replace with delete()
       $query_exp = "DELETE
                     FROM `glpi_crontasklogs`
                     WHERE `crontasks_id` = '$id'
                           AND UNIX_TIMESTAMP(date) < UNIX_TIMESTAMP()-$secs";
 
-      $DB->query($query_exp);
-      return $DB->affected_rows();
+      $result = $DB->query($query_exp);
+      return $result->rowCount();
    }